Did major progress on the dance:
LedControl: added functions to turn on and off all LEDs and changed randomLed function to pick random lights instead of flashing them.

Armcontrol: inverted the arms to the correct identification, and removed the delay before the arm moves as it was causing issues.

Dance: reworked how some moves worked, and now it works perfectly until move 20.